Create Biography
Mehdi Mohagheghi is a music content producer and held the PhD level of music and content of musical instruments training.
He hel the second place of the TopTen the best sales of the TopTen of musical instruments training content produced by Mehdi Mohagheghi which are being sold concerning several certified company and sites which are published by RGB in Spain in 2009 and so many positive opinion on this amazing visual and standard notation of guitar specially flamenco genre were passed on under the name of Progressive Studies by RGB in Spain for the second place of the TopTen 2010. He holds a badge of honour from the cultural ministry of Iran and so many positive feedback rating by the UNESCO and his fans.